Removed Copy-Frameworks build phase
This was causing the dependent frameworks to be copied with all archs. Including [x86_64, i386] that causes an issue when submitting to the App Store. See https://github.com/Carthage/Carthage/issues/416 and https://github.com/Carthage/Carthage/issues/353 for more info.

The tests still seems to work for me without this. This will require that the user of this framework provide the ReactiveCocoa och Result framework there selfes. But should fix the App Store submit problem.